First commit for the Spring reference guide template
I created a (hopefully) complete template for a Spring reference guide. It inclues a pom.xml file for building with Maven, Tocbot for a nice collapsing table of contents, and the spring.css file (to which we're steadily switching all the Spring reference guides).

This project is meant to serve as a starting place for new Spring reference guides and as a template for existing Spring reference guides that want to pick up the collapsing table of contents and the correct Spring styles and that want to get all four outputs (multiple HTML files, a single HTML file, PDF, and Epub).