Many changes to get epub to work
The Asciidoctor-epub3 generator is still in Alpha, and it's both brittle and missing some features. I had to tinker quite a lot to get everything to work and to come up with adequate substitutes for some of the things we got through the Docbook tool chain. I had to make a preamble.adoc file to capture the information in index-docinfo.xml, which would go before the table of contents in the PDF file but ends up after it now. It disappears altogether in epub, which is why I had to make the preamble.adoc file. It looks the same in the HTML outputs. Also, the ToC for the epub is in the Epub reader, not at the top of the document.